Updated swipl and fixed powershell commands

+++ scripts/prepareTestDirectory.ps1
@@ -1,4 +1,3 @@
-Remove-Item -Recurse src\test-temp
 mkdir src\test-temp
 cd src\test-temp
 mkdir agents
@@ -6,18 +5,26 @@ mkdir languages
 mkdir languages\test-language
 cp -r ..\tests\test-language\build languages\test-language\build
 
-Remove-Item hc.exe
+if (Test-Path hc.exe) {
+  Remove-Item hc.exe
+}
 $HcPath = Get-Command hc.exe | Select-Object -ExpandProperty Definition
 Copy-Item $HcPath -Destination hc.exe
 
-Remove-Item holochain.exe
+if (Test-Path holochain.exe) {
+  Remove-Item holochain.exe
+}
 $HoloPath = Get-Command holochain.exe | Select-Object -ExpandProperty Definition
 Copy-Item $HoloPath -Destination holochain.exe
 
-Remove-Item lair-keystore.exe
+if (Test-Path lair-keystore.exe) {
+  Remove-Item lair-keystore.exe
+}
 $LkPath = Get-Command lair-keystore.exe | Select-Object -ExpandProperty Definition
 Copy-Item $LkPath -Destination lair-keystore.exe
 
-Remove-Item swipl.exe
+if (Test-Path swipl.exe) {
+  Remove-Item swipl.exe
+}
 $LkPath = Get-Command swipl.exe | Select-Object -ExpandProperty Definition
 Copy-Item $LkPath -Destination swipl.exe
